web-app-project/backend/requirements.txt
Flask
requests
Adafruit-DHT
RPi.GPIO
adafruit-circuitpython-dht
flask-cors
google-generativeai